pyproject.toml
[tool.poetry]
name = "projeto-spotify"
version = "0.1.0"
description = ""
authors = ["Roberta Ferreira Aguilar <roberta.aguilar@kobel.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
spotipy = "^2.24.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
numpy = "^2.1.3"
matplotlib = "^3.9.3"
seaborn = "^0.13.2"
plotly = "^5.24.1"
streamlit = "^1.40.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
